fix type issues in ode_test

+++ jax/experimental/ode.py
@@ -265,13 +265,13 @@ def _odeint_rev(func, rtol, atol, mxstep, res, g):
 
   y_bar = g[-1]
   ts_bar = []
-  t0_bar = 0.
+  t0_bar = jnp.float_(0.)
 
   def scan_fun(carry, i):
     y_bar, t0_bar, args_bar = carry
     # Compute effect of moving measurement time
     t_bar = jnp.dot(func(ys[i], ts[i], *args), g[i])
-    t0_bar = t0_bar - t_bar
+    t0_bar = jnp.float_(t0_bar - t_bar)
     # Run augmented system backwards to previous observation
     _, y_bar, t0_bar, args_bar = odeint(
         aug_dynamics, (ys[i], y_bar, t0_bar, args_bar),
